requirements.txt
Flask==2.3.3
supabase==2.3.4
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
mistralai==1.0.1 